backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
pydantic==2.6.3
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0.post1 